requirements.txt
ansicon==1.89.0
arrow==1.2.3
asgiref==3.7.2
backports.zoneinfo==0.2.1
blessed==1.20.0
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.19
django-crispy-forms==2.0
django-currentuser==0.5.3
django-filter==23.2
django-import-export==3.2.0
django-mathfilters==1.0.0
django-picklefield==3.1
django-q==1.3.9
django-simple-history==3.3.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
importlib-metadata==6.6.0
jinxed==1.2.0
Markdown==3.4.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==3.5.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
